feat: Implement `ChooseFileButton` for `ChatInput`
