+++ src/containers/RoomContainer/useChatInput.ts
@@ -1,10 +1,7 @@
 import useConnection from "@/hooks/matrix/useConnection"
 import useDebounced from "@/hooks/util/useDebounced"
-import {sendImageMessageFromFile} from "@/utils/util"
 import {MsgType} from "matrix-js-sdk"
-import {useEffect, useMemo, useState} from "react"
-import {useFilePicker} from "use-file-picker"
-import {type ImageModalPreviewProps} from "./ImageModalPreview"
+import {useEffect, useState} from "react"
 
 type UseChatInputReturnType = {
   messageText: string
@@ -12,41 +9,12 @@ type UseChatInputReturnType = {
   isDisabled: boolean
   isInputDisabled: boolean
   sendTextMessage: (text: string) => Promise<void>
-  openFilePicker: () => void
-  imagePreviewProps: ImageModalPreviewProps | undefined
 }
 
 const useChatInput = (roomId: string): UseChatInputReturnType => {
   const {client} = useConnection()
   const [messageText, setMessageText] = useState("")
   const debouncedText = useDebounced(messageText, 500)
-
-  const {openFilePicker, filesContent, clear} = useFilePicker({
-    accept: "image/*",
-    multiple: false,
-    readAs: "DataURL",
-  })
-
-  const imagePreviewProps = useMemo(() => {
-    if (filesContent.length <= 0) {
-      return
-    }
-
-    const imageModalPreviewProps: ImageModalPreviewProps = {
-      imageName: filesContent[0].name,
-      imageUrl: filesContent[0].content,
-      onClear: clear,
-      onSendImage() {
-        void sendImageMessageFromFile(filesContent[0], client, roomId).then(
-          () => {
-            clear()
-          }
-        )
-      },
-    }
-
-    return imageModalPreviewProps
-  }, [clear, client, filesContent, roomId])
 
   const sendTextMessage = async (text: string): Promise<void> => {
     if (client === null || text.length === 0) {
@@ -81,8 +49,6 @@ const useChatInput = (roomId: string): UseChatInputReturnType => {
     isDisabled: client === null || messageText.length === 0,
     isInputDisabled: client === null,
     sendTextMessage,
-    openFilePicker,
-    imagePreviewProps,
   }
 }
